Autoformat, Add runworkflow comments, logtransform stub, savepath edits
Pycharm autoformat.
Add comments in run workflow.
Add stub for logtransform, and add to workflow
Change the file around so that the workflow can output to a main folder for the filter options and subfolders for clustering outputs.
Move the 11-25 data test from main() into separate functions.
It might not be fully complete yet? I'm not going to push until tomorrow.